requirements.txt
numpy==1.19.5
matplotlib==3.4.1
plotly==4.14.3
Django==3.1.3
pandas==1.2.2
django-picklefield==3.0.1
scikit_learn==0.24.1
